Revert "Merge pull request #1 from pakinkantawong/lab8-2-ipc"
This reverts commit 23f933c14e0751ae7abad667fc471f1a5d6e9b76, reversing
changes made to 632233a8e5c911c543c02398e160cf460b535dd5.
